+++ docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the `tkww-assistant` extension will be documented in this file.
 
+## [0.2.0] - 2020-11-02
+### Added
+- Added setting to toggle convert file on save.
+- Added `Convert Workspace to Union Variables` command to convert all files in workspace.
+- Added `Toggle Convert File on Save` command to toggle convert file on save.
+
 ## [0.1.0] - 2020-11-01
 ### Changed
 - Change codebase to use abstract syntax tree for more stability